+++ modules/jaxb-xml-binding/geotk-xml-gml/src/main/java/org/geotoolkit/gml/xml/v321/AffinePlacementType.java
@@ -18,7 +18,6 @@
 
 package org.geotoolkit.gml.xml.v321;
 
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 import javax.xml.bind.annotation.XmlAccessType;
@@ -65,10 +64,10 @@ public class AffinePlacementType {
     private List<VectorType> refDirection;
     @XmlElement(required = true)
     @XmlSchemaType(name = "positiveInteger")
-    private BigInteger inDimension;
+    private Integer inDimension;
     @XmlElement(required = true)
     @XmlSchemaType(name = "positiveInteger")
-    private BigInteger outDimension;
+    private Integer outDimension;
 
     /**
      * Gets the value of the location property.
@@ -128,10 +127,10 @@ public class AffinePlacementType {
      * 
      * @return
      *     possible object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public BigInteger getInDimension() {
+    public Integer getInDimension() {
         return inDimension;
     }
 
@@ -140,10 +139,10 @@ public class AffinePlacementType {
      * 
      * @param value
      *     allowed object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public void setInDimension(BigInteger value) {
+    public void setInDimension(Integer value) {
         this.inDimension = value;
     }
 
@@ -152,10 +151,10 @@ public class AffinePlacementType {
      * 
      * @return
      *     possible object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public BigInteger getOutDimension() {
+    public Integer getOutDimension() {
         return outDimension;
     }
 
@@ -164,10 +163,10 @@ public class AffinePlacementType {
      * 
      * @param value
      *     allowed object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public void setOutDimension(BigInteger value) {
+    public void setOutDimension(Integer value) {
         this.outDimension = value;
     }
