+++ modules/jaxb-xml-binding/geotk-xml-gml/src/main/java/org/geotoolkit/gml/xml/v321/ArcByCenterPointType.java
@@ -18,7 +18,6 @@
 
 package org.geotoolkit.gml.xml.v321;
 
-import java.math.BigInteger;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlAttribute;
@@ -89,7 +88,7 @@ public class ArcByCenterPointType
     @XmlAttribute
     private CurveInterpolationType interpolation;
     @XmlAttribute(required = true)
-    private BigInteger numArc;
+    private Integer numArc;
 
     /**
      * Gets the value of the pos property.
@@ -316,12 +315,12 @@ public class ArcByCenterPointType
      * 
      * @return
      *     possible object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public BigInteger getNumArc() {
+    public Integer getNumArc() {
         if (numArc == null) {
-            return new BigInteger("1");
+            return new Integer("1");
         } else {
             return numArc;
         }
@@ -332,10 +331,10 @@ public class ArcByCenterPointType
      * 
      * @param value
      *     allowed object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public void setNumArc(BigInteger value) {
+    public void setNumArc(Integer value) {
         this.numArc = value;
     }
