+++ modules/jaxb-xml-binding/geotk-xml-gml/src/main/java/org/geotoolkit/gml/xml/v321/KnotType.java
@@ -18,7 +18,6 @@
 
 package org.geotoolkit.gml.xml.v321;
 
-import java.math.BigInteger;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
@@ -58,7 +57,7 @@ public class KnotType {
     private double value;
     @XmlElement(required = true)
     @XmlSchemaType(name = "nonNegativeInteger")
-    private BigInteger multiplicity;
+    private Integer multiplicity;
     private double weight;
 
     /**
@@ -82,10 +81,10 @@ public class KnotType {
      * 
      * @return
      *     possible object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public BigInteger getMultiplicity() {
+    public Integer getMultiplicity() {
         return multiplicity;
     }
 
@@ -94,10 +93,10 @@ public class KnotType {
      * 
      * @param value
      *     allowed object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public void setMultiplicity(BigInteger value) {
+    public void setMultiplicity(Integer value) {
         this.multiplicity = value;
     }
